+++ src/constants.ts
@@ -1,3 +1,1 @@
 export const VIEW_TYPE_WROT = "wrot-view";
-
-export const WR_LANG = "wr";